backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
starlette==0.27.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client for API calls
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Payment processing
stripe==7.8.0

# Email service
sendgrid==6.11.0

# CORS middleware (included with FastAPI)
# Additional utilities
python-dateutil==2.8.2

# Input validation and sanitization
bleach==6.1.0
